♻️ Refactors projects plugin sub-modules as controller-service-repository (#4389)
