--- api/specs/webserver/openapi-node-v0.0.1.yaml
@@ -1,110 +1,0 @@
-paths:
-  /nodes/{nodeInstanceUUID}/outputUi/{outputKey}:
-    get:
-      tags:
-        - node
-      description:
-        get a json description of the ui for presenting the output within the mainUi
-        and a list of open api json schema objects describing the possible
-        json payloads and responses for the api calls available at this endpoint
-      operationId: get_node_output_ui
-      parameters:
-        - $ref: "#/components/parameters/nodeInstanceUUID"
-        - $ref: "#/components/parameters/outputKey"
-      responses:
-        "200":
-          description: Service Information
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/outputUi"
-        default:
-          description: Unexpected error
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ErrorEnveloped"
-
-  /nodes/{nodeInstanceUUID}/outputUi/{outputKey}/{apiCall}:
-    post:
-      tags:
-        - node
-      summary: send data back to the output api ... protocol depends on the definition
-      operationId: send_to_node_output_api
-      parameters:
-        - $ref: "#/components/parameters/nodeInstanceUUID"
-        - $ref: "#/components/parameters/outputKey"
-        - $ref: "#/components/parameters/apiCall"
-      requestBody:
-        content:
-          application/json:
-            schema:
-              oneOf:
-                - $ref: "../common/schemas/node-output-list-api-v0.0.1.yaml#/properties/getItemList/properties/request"
-                - $ref: "../common/schemas/node-output-list-api-v0.0.1.yaml#/properties/getItem/properties/request"
-                - $ref: "../common/schemas/node-output-tree-api-v0.0.1.yaml#/properties/getItemList/properties/request"
-                - $ref: "../common/schemas/node-output-tree-api-v0.0.1.yaml#/properties/getItem/properties/request"
-      responses:
-        default:
-          description: node type specific api call according to the node type presented
-          content:
-            application/json:
-              schema:
-                oneOf:
-                  - $ref: "../common/schemas/node-output-list-api-v0.0.1.yaml#/properties/getItemList/properties/response"
-                  - $ref: "../common/schemas/node-output-list-api-v0.0.1.yaml#/properties/getItem/properties/response"
-                  - $ref: "../common/schemas/node-output-tree-api-v0.0.1.yaml#/properties/getItemList/properties/response"
-                  - $ref: "../common/schemas/node-output-tree-api-v0.0.1.yaml#/properties/getItem/properties/response"
-
-  /nodes/{nodeInstanceUUID}/iframe:
-    get:
-      tags:
-        - node
-      summary: entry point for iframe interaction with the node. This relies on the reverse proxy code.
-      operationId: get_node_output_iframe
-      parameters:
-        - $ref: "#/components/parameters/nodeInstanceUUID"
-      responses:
-        default:
-          description: "any response appropriate in the iframe context"
-
-components:
-  parameters:
-    nodeInstanceUUID:
-      in: path
-      name: nodeInstanceUUID
-      required: true
-      schema:
-        type: string
-    outputKey:
-      in: path
-      name: outputKey
-      required: true
-      schema:
-        type: string
-    apiCall:
-      in: path
-      name: apiCall
-      required: true
-      schema:
-        type: string
-  schemas:
-    outputUi:
-      type: object
-      properties:
-        plugin:
-          type: string
-        config:
-          type: object
-    outputApiCall:
-      type: object
-      properties:
-        call:
-          type: string
-        request:
-          type: object
-    outputApiResponse:
-      type: object
-
-    ErrorEnveloped:
-      $ref: "../common/schemas/error.yaml#/components/schemas/ErrorEnveloped"
