+++ api/specs/webserver/openapi.yaml
@@ -1,7 +1,7 @@
 openapi: 3.0.0
 info:
   title: "osparc-simcore web API"
-  version: 0.22.0
+  version: 0.23.0
   description: "API designed for the front-end app"
   contact:
     name: IT'IS Foundation
@@ -32,7 +32,6 @@ tags:
   - name: configuration
   - name: maintenance
   - name: nih-sparc
-  - name: node
   - name: project
   - name: publication
   - name: repository
@@ -256,15 +255,6 @@ paths:
 
   /projects/{project_id}/metadata/ports:
     $ref: "./openapi-projects-ports.yaml#/paths/~1projects~1{project_id}~1metadata~1ports"
-
-  /nodes/{nodeInstanceUUID}/outputUi/{outputKey}:
-    $ref: "./openapi-node-v0.0.1.yaml#/paths/~1nodes~1{nodeInstanceUUID}~1outputUi~1{outputKey}"
-
-  /nodes/{nodeInstanceUUID}/outputUi/{outputKey}/{apiCall}:
-    $ref: "./openapi-node-v0.0.1.yaml#/paths/~1nodes~1{nodeInstanceUUID}~1outputUi~1{outputKey}~1{apiCall}"
-
-  /nodes/{nodeInstanceUUID}/iframe:
-    $ref: "./openapi-node-v0.0.1.yaml#/paths/~1nodes~1{nodeInstanceUUID}~1iframe"
 
   /projects/{study_uuid}/tags/{tag_id}:
     $ref: "./openapi-projects.yaml#/paths/~1projects~1{study_uuid}~1tags~1{tag_id}"
