+++ services/web/server/VERSION
@@ -1,1 +1,1 @@
-0.22.0
+0.23.0
