+++ services/web/server/setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.22.0
+current_version = 0.23.0
 commit = True
 message = services/webserver api version: {current_version} → {new_version}
 tag = False
