+++ services/web/server/src/simcore_service_webserver/projects/_nodes_api.py
@@ -1,24 +1,71 @@
+import datetime
 import logging
 import mimetypes
 import urllib.parse
+from typing import Final
 
 from aiohttp import web
 from aiohttp.client import ClientError
+from models_library.projects import ProjectID
 from models_library.projects_nodes import Node, NodeID
 from models_library.projects_nodes_io import SimCoreFileLink
+from models_library.users import UserID
 from pydantic import (
     BaseModel,
     Field,
     HttpUrl,
+    NonNegativeFloat,
+    NonNegativeInt,
     ValidationError,
     parse_obj_as,
     root_validator,
 )
 
 from .._constants import APP_SETTINGS_KEY, RQT_USERID_KEY
+from ..application_settings import get_settings
 from ..storage.api import get_download_link
+from .exceptions import ProjectStartsTooManyDynamicNodes
 
 _logger = logging.getLogger(__name__)
+_NODE_START_INTERVAL_S: Final[datetime.timedelta] = datetime.timedelta(seconds=15)
+
+
+def get_service_start_lock_key(user_id: UserID, project_uuid: ProjectID) -> str:
+    return f"lock_service_start_limit.{user_id}.{project_uuid}"
+
+
+def check_num_service_per_projects_limit(
+    app: web.Application,
+    number_of_services: int,
+    user_id: UserID,
+    project_uuid: ProjectID,
+) -> None:
+    """
+    raises ProjectStartsTooManyDynamicNodes if the user cannot start more services
+    """
+    project_settings = get_settings(app).WEBSERVER_PROJECTS
+    assert project_settings  # nosec
+    if project_settings.PROJECTS_MAX_NUM_RUNNING_DYNAMIC_NODES > 0 and (
+        number_of_services >= project_settings.PROJECTS_MAX_NUM_RUNNING_DYNAMIC_NODES
+    ):
+        raise ProjectStartsTooManyDynamicNodes(
+            user_id=user_id, project_uuid=project_uuid
+        )
+
+
+def get_total_project_dynamic_nodes_creation_interval(
+    max_nodes: NonNegativeInt,
+) -> NonNegativeFloat:
+    """
+    Estimated amount of time for all project node creation requests to be sent to the
+    director-v2. Note: these calls are sent one after the other.
+    """
+    return max_nodes * _NODE_START_INTERVAL_S.total_seconds()
+
+
+#
+# PREVIEWS
+#
 
 
 class NodeScreenshot(BaseModel):
