+++ services/web/server/src/simcore_service_webserver/projects/_permalink_api.py
