+++ services/web/server/src/simcore_service_webserver/projects/api.py
@@ -1,8 +1,8 @@
 # NOTE: we will slowly move heere projects_api.py
 
 
-from ._permalink import ProjectPermalink
-from ._permalink import register_factory as register_permalink_factory
+from ._permalink_api import ProjectPermalink
+from ._permalink_api import register_factory as register_permalink_factory
 
 __all__: tuple[str, ...] = (
     "register_permalink_factory",
