+++ services/web/server/src/simcore_service_webserver/projects/plugin.py
@@ -18,9 +18,9 @@ from . import (
     _comments_handlers,
     _handlers,
     _handlers_crud,
-    _handlers_project_nodes,
-    _handlers_project_ports,
-    _handlers_project_tags,
+    _nodes_handlers,
+    _ports_handlers,
+    _tags_handlers,
 )
 from ._observer import setup_project_observer_events
 from ._projects_access import setup_projects_access
@@ -70,15 +70,15 @@ def setup_projects(app: web.Application) -> bool:
 
     app.router.add_routes(_handlers.routes)
     app.router.add_routes(_handlers_crud.routes)
-    app.router.add_routes(_handlers_project_ports.routes)
     app.router.add_routes(_comments_handlers.routes)
+    app.router.add_routes(_ports_handlers.routes)
 
     app.router.add_routes(
         _create_routes(
             "project",
             specs,
-            _handlers_project_nodes,
-            _handlers_project_tags,
+            _nodes_handlers,
+            _tags_handlers,
         )
     )
