+++ services/web/server/tests/unit/with_dbs/02/test_projects_handlers_crud.py
@@ -29,7 +29,7 @@ from simcore_postgres_database.models.projects_to_products import projects_to_pr
 from simcore_service_webserver._constants import X_PRODUCT_NAME_HEADER
 from simcore_service_webserver._meta import api_version_prefix
 from simcore_service_webserver.db.models import UserRole
-from simcore_service_webserver.projects._permalink import ProjectPermalink
+from simcore_service_webserver.projects._permalink_api import ProjectPermalink
 from simcore_service_webserver.projects.models import ProjectDict
 from simcore_service_webserver.utils import to_datetime
 from yarl import URL
