+++ services/web/server/tests/unit/with_dbs/02/test_projects_ports_handlers.py
@@ -19,14 +19,14 @@ from pytest_simcore.helpers.utils_assert import assert_status
 from pytest_simcore.helpers.utils_login import UserInfoDict
 from simcore_service_webserver._meta import API_VTAG as VX
 from simcore_service_webserver.db.models import UserRole
-from simcore_service_webserver.projects import _handlers_project_ports
+from simcore_service_webserver.projects import _ports_handlers
 from simcore_service_webserver.projects.models import ProjectDict
 from yarl import URL
 
 
 @pytest.mark.parametrize(
     "route",
-    _handlers_project_ports.routes,
+    _ports_handlers.routes,
     ids=lambda r: f"{r.method.upper()} {r.path}",
 )
 def test_route_against_openapi_specs(route, openapi_specs: OpenApiSpecs):
